Add examples in `platform_pki_2_electric_boogaloo`: Update `utils/mod.rs`
- Make clear content passed as param should be in base64
- Fix content passed as base64 being used as-is instead of being decoded
- Remove `CertificateOrRef` to only allow specifying a certificate by its fingerprint.
  This is because the new API doesn't accept certificate as PEM/DER file (it is designed
  to work with the certificate store instead).
- Remove `display_x509_raw_name` & `display_cert` only used in `list_certificate` example so moved there
